requirements.txt
flask==3.0.2
flask-cors==4.0.0
streamlit==1.32.2
pyjwt==2.8.0
requests==2.31.0
